Use GitlabUploadIT as another integration test
